LayoutTests/http/wpt/content-security-policy/sandbox-manifest-blocked-cross-origin.sub-expected.txt

http://localhost:8800/WebKit/content-security-policy/sandbox-manifest-blocked-cross-origin.sub.html - didFinishLoading
CONSOLE MESSAGE: Origin null is not allowed by Access-Control-Allow-Origin. Status code: 200
http://not-web-platform.test:8800/WebKit/content-security-policy/manifest.json - didFailLoadingWithError: <NSError domain , code 0, failing URL "http://not-web-platform.test:8800/WebKit/content-security-policy/manifest.json">
CONSOLE MESSAGE: Fetched manifest: http://not-web-platform.test:8800/WebKit/content-security-policy/manifest.json
